Crate wasmer_pack_testing

Utilities for testing bindings generated by Wasmer Pack.

Typical.

let env = TestEnvironment::for_crate("./path/to/Cargo.toml", "/tmp")?;
env.python("./my_tests.py")?;
env.javascript("./my_test.js")?;
env.typescript("./my_tests.ts")?;

Under the hood, this will use cargo wapm to compile a Rust crate to WebAssembly and turn it into a WAPM package.

You can tell it to run test scripts written in various languages.

The TestEnvironment::python() method will create a Virtual Environment in the script’s directory and install the generated Python library. The provided test script will then be run in that environment using py.test.

The TestEnvironment::javascript() and TestEnvironment::typescript() methods will generate JavaScript bindings for the Rust crate and use yarn link to add them as a dependency. From there, the test script will be run using Jest.

Structs

TestEnvironment

Enums

CommandFailed

Language

LoadError

TestFailure

Functions

autodiscover

compile_rust_to_wapm_package

generate_bindings